// branchClient/src/app/CDBL/models/boClient.ts
export interface IBoClient {
    rcode: string | null;
    slno: number | null;
    boid: string | null;
    botype: string | null;
    bocat: string | null;
    fhName: string | null;
    shName: string | null;
    thName: string | null;
    cpName: string | null;
    sex: string | null;
    dob: string | null;
    regNum: string | null;
    father: string | null;
    mother: string | null;
    occupation: string | null;
    residency: string | null;
    nationality: string | null;
    add1: string | null;
    add2: string | null;
    add3: string | null;
    city: string | null;
    state: string | null;
    country: string | null;
    pCode: string | null;
    mobile: string | null;
    mobile1: string | null;
    email: string | null;
    fax: string | null;
    scc: string | null;
    fhShort: string | null;
    shShort: string | null;
    thShort: string | null;
    passNum: string | null;
    passIDate: string | null;
    passExDate: string | null;
    passIPlace: string | null;
    bName: string | null;
    bBranch: string | null;
    bAccNum: string | null;
    eftStatus: string | null;
    taxStatus: string | null;
    tin: string | null;
    eID: string | null;
    tID: string | null;
    routing: string | null;
    bCode: string | null;
    intBAcctNum: string | null;
    swift: string | null;
    setupDate: string | null;
    boStatus: string | null;
    closeDate: string | null;
    upFileDate: string | null;
    lastUpdate: string | null;
    nID1: string | null;
    nID2: string | null;
    nID3: string | null;
}